[core] fix MBR cleanup regression
* Commit 9dc045a7017cfd3ac503ced89a3a1c6604dd5d67 introduced a regression on account that we didn't set the
  file pointer to 0 before clearing the disk.
* This leads to the MBR not being properly cleared, with the result that Windows may in turn
  produce errors when trying to repartition the disk.
* Fix this by making sure we do invoke SetFilePointerEx() before calling WriteFileWithRetry().
* Also set rufus-next to 3.17
* Also fix a MinGW warning

+++ src/stdfn.c
@@ -310,7 +310,7 @@ static const char* GetEdition(DWORD ProductType)
 	case 0x000000BD: return "Lite";
 	case 0xABCDABCD: return "(Unlicensed)";
 	default:
-		static_sprintf(unknown_edition_str, "(Unknown Edition 0x%02X)", ProductType);
+		static_sprintf(unknown_edition_str, "(Unknown Edition 0x%02X)", (uint32_t)ProductType);
 		return unknown_edition_str;
 	}
 }
